+++ packages/calling-component-bindings/src/hooks/usePropsFor.ts
@@ -9,7 +9,7 @@ import {
   ParticipantList,
   ScreenShareButton,
   VideoGallery
-} from 'react-components';
+} from '@internal/react-components';
 import {
   cameraButtonSelector,
   microphoneButtonSelector,
@@ -24,7 +24,7 @@ import { useSelector } from './useSelector';
 import { Common } from '@internal/acs-ui-common';
 import { AreEqual } from '@internal/acs-ui-common';
 import { DefaultCallingHandlers } from '../handlers/createHandlers';
-import { ParticipantsButton } from 'react-components';
+import { ParticipantsButton } from '@internal/react-components';
 
 export const usePropsFor = <Component extends (props: any) => JSX.Element>(
   component: Component