+++ packages/chat-component-bindings/src/chatParticipantListSelector.ts
@@ -5,7 +5,7 @@ import { getUserId, getDisplayName, getParticipants } from './baseSelectors';
 import * as reselect from 'reselect';
 import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
 import { ChatParticipant } from '@azure/communication-chat';
-import { CommunicationParticipant } from 'react-components';
+import { CommunicationParticipant } from '@internal/react-components';
 
 const convertChatParticipantsToCommunicationParticipants = (
   chatParticipants: ChatParticipant[]