+++ packages/storybook/stories/ErrorBar/ErrorBar.stories.REMOVED.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 /**
- * THIS COMPONENT HAS BEEN REMOVED FROM REACT-COMPONENTS PACKAGE.
+ * THIS COMPONENT HAS BEEN REMOVED FROM @internal/react-compoNENTS PACKAGE.
  * AS SUCH THIS STORY HAS BEEN MARKED 'REMOVED' HOWEVER MAY BE RETURNED
  * WHEN THE COMPOSITE ERROR HANDLING STORY HAS BEEN COMPLETED.
  */
@@ -11,7 +11,7 @@ import { Canvas, Description, Heading, Props, Source, Title } from '@storybook/a
 import { text, select } from '@storybook/addon-knobs';
 import { Meta } from '@storybook/react/types-6-0';
 import React, { useState } from 'react';
-import { CommunicationUiErrorSeverity, ErrorBar as ErrorBarComponent } from 'react-composites';
+import { CommunicationUiErrorSeverity, ErrorBar as ErrorBarComponent } from '@internal/react-composites';
 
 import { COMPONENT_FOLDER_PREFIX } from '../constants';
 import { ErrorBarExample } from './snippets/ErrorBar.snippet';