+++ packages/storybook/stories/Examples/IncomingCallAlerts/IncomingCallModal.tsx
@@ -3,7 +3,7 @@
 
 import { text, files, boolean } from '@storybook/addon-knobs';
 import React from 'react';
-import { IncomingCallModal as IncomingCallModalComponent } from 'react-composites';
+import { IncomingCallModal as IncomingCallModalComponent } from '@internal/react-composites';
 import { useVideoStreams } from '../../utils';
 
 export const IncomingCallModal: () => JSX.Element = () => {